Hopefully maybe fix a cryptic Beogh conversion crash
It's not entirely clear to me the sequence of events required to trigger
the crash in a real game, though at least two people have. It seems to
(maybe) require returning to a level with no monsters on it, which still
somehow has the LSTATE_BEOGH flag set, while the player is injured enough
to be able to convert to Beogh. Though it's unclear to me how that flag
can *be* set with no monsters on the floor (it also probably only happens
on webtiles, due to codepaths involving milestones being examined on floor
transition.)

But at any rate, env.max_mon_index should probably never be -1, even if
it's unclear *exactly* why that is resulting in this crash via
monster_near_iterator.