Fix webtiles desync after spectator joins
Webtiles servers usually only send the delta from the last frame when
sendint a frame, however, when a spectator joins a full frame needs to
be sent. When sending a full frame, things (e.g. overlays, background,
foreground) that have their default value aren't sent and the clients
which use default values for anything they are missing. The server can
have a frame that hasn't been sent to the clients yet when a spectator
joins, which means that if a tile has a non-default state (e.g. a fire
ball overlay over it) and this is sent to the clients and then this
state returns to default (e.g. the fire ball overlay disapears) but
is hasn't been sent to the client yet when a spectactor joins causing
a full frame to be sent, this now default state won't be send to the
clients who still believe it have have its previous state (with the
exception of the specator that just joined). To fix this, when sending
a full frame send the state of a tile to the clients when it is
non-default or different from previous.